feat: 사진 기록, 경로 기록 컴포넌트로 구분
